Move `setup_shim()` to `setup` module
